chore(xlsx): extract chart series host into chart/series
Move parseSeries / parseSmooth / parseInvertIfNegative / parseExplosion /
parseMarker / parseSeriesName / parseSeriesColor / parseSeriesStroke
(reader), buildSeries / buildSeriesSpPr / buildSeriesMarker /
clampExplosion / clampMarkerSize / normalizeMarkerSymbol /
normalizeDashStyle (writer), and resolveSmooth / resolveInvertIfNegative /
resolveExplosion / resolveMarker / cloneMarker / resolveStroke /
cloneStroke / resolveShowLineMarkers (clone) into chart/series.ts.

Move the writer-side SeriesOptions interface, EXPLOSION_MAX,
VALID_MARKER_SYMBOLS, MARKER_SIZE_MIN/MAX, and the qualifyRef /
quoteSheetName helpers out of chart-writer.ts into chart/series.ts
because they are exclusively used by the moved series functions.

The clone-side buildSeriesFromSource / mergeSeries stay in
chart-clone.ts because they depend on the local applyOverride helper
and the public CloneChartSeriesOverride interface, both of which would
require an additional cross-module move; they delegate to the per-field
resolvers / cloners exported here.

Pure relocation — JSDoc preserved verbatim, no behavior change. All
7362 tests still pass; typecheck clean.
