+++ src/xlsx/chart-writer.ts
@@ -113,6 +113,7 @@ import {
   buildChartLevelDataLabels,
   buildSeriesDataLabels,
 } from "./chart/dataLabels";
+import { buildSeries } from "./chart/series";
 
 // ── Namespaces ───────────────────────────────────────────────────────
 
@@ -3077,29 +3078,6 @@ function clampHoleSize(value: number | undefined): number {
   return rounded;
 }
 
-const EXPLOSION_MAX = 400;
-
-/**
- * Normalize {@link ChartSeries.explosion} for emission inside
- * `<c:explosion val=".."/>` on a pie / doughnut series.
- *
- * The OOXML schema (`CT_UnsignedInt`) accepts any non-negative integer,
- * but Excel's UI only exposes 0..400% — values outside that band render
- * but trigger Excel's repair dialog. Clamp to the UI band on the way
- * out so a round-trip stays inside the range Excel will render.
- *
- * Returns `undefined` for the default `0` (and any negative / non-finite
- * input) so the writer can elide the element entirely; absence and `0`
- * round-trip identically through the parser's collapse logic.
- */
-function clampExplosion(value: number | undefined): number | undefined {
-  if (value === undefined || !Number.isFinite(value)) return undefined;
-  const rounded = Math.round(value);
-  if (rounded <= 0) return undefined;
-  if (rounded > EXPLOSION_MAX) return EXPLOSION_MAX;
-  return rounded;
-}
-
 // ── Scatter ──────────────────────────────────────────────────────────
 
 function buildScatterChart(chart: SheetChart, sheetName: string): string {
@@ -3635,357 +3613,6 @@ function normalizeAxisTitleFontFamily(value: string | undefined): string | undef
   const trimmed = value.trim();
   if (trimmed.length === 0) return undefined;
   return trimmed;
-}
-
-// ── Series ───────────────────────────────────────────────────────────
-
-interface SeriesOptions {
-  smooth?: boolean;
-  /**
-   * Owning chart's family. Used to scope-guard schema-restricted
-   * data-label flags such as `<c:showLeaderLines>` (pie / doughnut
-   * only) so a templated chart's pin never leaks onto a chart family
-   * Excel's strict validator would reject. Required so the per-family
-   * caller cannot forget to thread the type through; every chart
-   * builder passes `chart.type` directly.
-   */
-  chartType: WriteChartKind;
-  /**
-   * Chart-level data label defaults from {@link SheetChart.dataLabels}.
-   * Used when the series itself does not specify `dataLabels`. Series
-   * passing `dataLabels: false` always wins over this default.
-   */
-  dataLabels?: ChartDataLabels;
-  /**
-   * Per-series line stroke (dash pattern + width). Only meaningful for
-   * line / scatter series — every other family ignores the field. The
-   * OOXML schema places stroke styling inside `<c:spPr><a:ln>` which is
-   * shared with the series fill color, so the writer threads the
-   * stroke into the same `<c:spPr>` block whether or not a fill color
-   * is set.
-   */
-  stroke?: ChartLineStroke;
-  /**
-   * Per-series marker styling. Only meaningful for line / scatter
-   * series — every other family ignores the field. The OOXML schema
-   * places `<c:marker>` between `<c:spPr>` and `<c:dLbls>` on
-   * `CT_LineSer` / `CT_ScatterSer`, so the writer slots it there
-   * regardless of which fields are populated.
-   */
-  marker?: ChartMarker;
-  /**
-   * Per-series invert-if-negative flag. Only meaningful for bar /
-   * column series — every other family ignores the field. The OOXML
-   * schema places `<c:invertIfNegative>` between `<c:spPr>` and
-   * `<c:dLbls>` on `CT_BarSer` / `CT_Bar3DSer`, so the writer slots
-   * it there. The element is only emitted when the field resolves to
-   * `true` — `false` is the OOXML default and absence round-trips
-   * identically.
-   */
-  invertIfNegative?: boolean;
-  /**
-   * Per-series slice explosion (percentage of the radius). Only
-   * meaningful for pie / doughnut series — every other family ignores
-   * the field. The OOXML schema places `<c:explosion>` between
-   * `<c:spPr>` and `<c:dPt>` / `<c:dLbls>` on `CT_PieSer`. The element
-   * is only emitted when the resolved value is `> 0` — `0` is the OOXML
-   * default and absence round-trips identically.
-   */
-  explosion?: number;
-}
-
-function buildSeries(
-  series: ChartSeries,
-  index: number,
-  sheetName: string,
-  numericCategories: boolean,
-  options: SeriesOptions,
-): string {
-  const children: string[] = [
-    xmlSelfClose("c:idx", { val: index }),
-    xmlSelfClose("c:order", { val: index }),
-  ];
-
-  if (series.name) {
-    // Literal series names go inside <c:tx><c:v>…</c:v></c:tx>. Excel
-    // also accepts <c:strRef> for cell-bound names; literals are the
-    // simpler shape and round-trip just as well.
-    children.push(
-      xmlElement("c:tx", undefined, [xmlElement("c:v", undefined, xmlEscape(series.name))]),
-    );
-  }
-
-  // Optional fill color and / or line stroke (line / scatter only emit
-  // `<a:ln>` width / dash from `options.stroke`; non-line callers leave
-  // `options.stroke` undefined so the field is silently dropped on
-  // every other chart family).
-  const spPr = buildSeriesSpPr(series.color, options?.stroke);
-  if (spPr) children.push(spPr);
-
-  // Marker — only line/scatter series honor `<c:marker>` per the OOXML
-  // schema (CT_LineSer / CT_ScatterSer). The element sits between
-  // `<c:spPr>` and `<c:dLbls>`; non-line/non-scatter callers leave
-  // `options.marker` undefined so the field is silently dropped on
-  // every other chart family.
-  const markerXml = buildSeriesMarker(options?.marker);
-  if (markerXml) children.push(markerXml);
-
-  // `<c:invertIfNegative>` — only bar / column (CT_BarSer /
-  // CT_Bar3DSer) series carry the element per the OOXML schema. It
-  // sits between `<c:spPr>` (and the bar-irrelevant `<c:marker>`
-  // slot, which is never populated for bar/column callers anyway)
-  // and `<c:dLbls>`. Non-bar callers leave `options.invertIfNegative`
-  // undefined so the field is silently dropped on every other chart
-  // family. Emit only when the resolved value is `true` — `false`
-  // matches the OOXML default and absence round-trips identically.
-  if (options?.invertIfNegative === true) {
-    children.push(xmlSelfClose("c:invertIfNegative", { val: 1 }));
-  }
-
-  // `<c:explosion>` — only pie / doughnut (CT_PieSer, shared across
-  // the pie family via `EG_PieSer`) series carry the element per the
-  // OOXML schema. It sits between `<c:spPr>` and `<c:dPt>` / `<c:dLbls>`.
-  // Non-pie callers leave `options.explosion` undefined so the field
-  // is silently dropped on every other chart family. Emit only when
-  // the resolved value is non-zero — `0` matches the OOXML default and
-  // absence round-trips identically.
-  const explosion = clampExplosion(options?.explosion);
-  if (explosion !== undefined) {
-    children.push(xmlSelfClose("c:explosion", { val: explosion }));
-  }
-
-  // Data labels — series-level override always wins over the chart-level
-  // default. `<c:dLbls>` sits between <c:spPr> and <c:cat>/<c:val> per
-  // the OOXML series schema (CT_BarSer, CT_LineSer, ...). The chart
-  // type threads through so the dLbls body can scope-guard the pie /
-  // doughnut-only `<c:showLeaderLines>` flag.
-  const seriesDLblsXml = buildSeriesDataLabels(
-    series.dataLabels,
-    options.dataLabels,
-    options.chartType,
-  );
-  if (seriesDLblsXml) children.push(seriesDLblsXml);
-
-  // Categories (skipped for pie when omitted; allowed for all)
-  if (series.categories) {
-    const ref = qualifyRef(series.categories, sheetName);
-    if (numericCategories) {
-      children.push(
-        xmlElement("c:xVal", undefined, [
-          xmlElement("c:numRef", undefined, [xmlElement("c:f", undefined, xmlEscape(ref))]),
-        ]),
-      );
-    } else {
-      children.push(
-        xmlElement("c:cat", undefined, [
-          xmlElement("c:strRef", undefined, [xmlElement("c:f", undefined, xmlEscape(ref))]),
-        ]),
-      );
-    }
-  }
-
-  // Values
-  const valuesRef = qualifyRef(series.values, sheetName);
-  if (numericCategories) {
-    children.push(
-      xmlElement("c:yVal", undefined, [
-        xmlElement("c:numRef", undefined, [xmlElement("c:f", undefined, xmlEscape(valuesRef))]),
-      ]),
-    );
-  } else {
-    children.push(
-      xmlElement("c:val", undefined, [
-        xmlElement("c:numRef", undefined, [xmlElement("c:f", undefined, xmlEscape(valuesRef))]),
-      ]),
-    );
-  }
-
-  if (options?.smooth !== undefined) {
-    children.push(xmlSelfClose("c:smooth", { val: options.smooth ? 1 : 0 }));
-  }
-
-  return xmlElement("c:ser", undefined, children);
-}
-
-// ── Stroke ───────────────────────────────────────────────────────────
-//
-// `EMU_PER_PT`, `VALID_DASH_STYLES`, `VALID_BORDER_DASHES`,
-// `clampStrokeWidthPt`, and `normalizeBorderDash` now live in
-// `./chart/shape.ts`. Imported at the top of this module so the chart-
-// space / plot-area / legend / title / axis-title / data-labels /
-// data-table writers all share the same EMU encoding and accept-or-
-// drop dash grammar.
-
-/**
- * Validate a dash style against `ST_PresetLineDashVal`. Returns
- * `undefined` for unrecognized values so the writer can elide
- * `<a:prstDash>` rather than emit a token Excel will reject.
- */
-function normalizeDashStyle(value: ChartLineDashStyle | undefined): ChartLineDashStyle | undefined {
-  if (value === undefined) return undefined;
-  return VALID_DASH_STYLES.has(value) ? value : undefined;
-}
-
-/**
- * Build the `<c:spPr>` element shared by series fill color and series
- * line stroke. Returns `undefined` when neither field carries any
- * meaningful settings — an empty `<c:spPr/>` collapses to the
- * inherited series-rotation default Excel picks anyway, so omitting it
- * keeps untouched chart XML byte-clean.
- *
- * The OOXML `<a:ln>` element accepts both a `w` attribute (stroke
- * width in EMU) and child elements `<a:solidFill>` / `<a:prstDash>` in
- * a fixed order. When a fill color is set, the stroke also renders the
- * same color (matching Excel's "Format Data Series → Fill" default
- * which paints the line in the fill color). Stroke metadata (dash and
- * width) layers on top without overriding the line color so a `color +
- * stroke` combo behaves like Excel's UI: the line picks up the fill
- * color and the dash / width override visibility-only attributes.
- */
-function buildSeriesSpPr(
-  rgbHex: string | undefined,
-  stroke: ChartLineStroke | undefined,
-): string | undefined {
-  const fillHex = rgbHex ? rgbHex.replace(/^#/, "").toUpperCase() : undefined;
-  const dash = normalizeDashStyle(stroke?.dash);
-  const widthPt = clampStrokeWidthPt(stroke?.width);
-
-  if (!fillHex && dash === undefined && widthPt === undefined) {
-    return undefined;
-  }
-
-  const spPrChildren: string[] = [];
-  if (fillHex) {
-    spPrChildren.push(
-      xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: fillHex })]),
-    );
-  }
-
-  // `<a:ln>` carries stroke metadata. Emit it whenever a fill color is
-  // set (so the connecting line picks up the same color, matching the
-  // legacy behavior) or whenever stroke width / dash is configured.
-  if (fillHex || dash !== undefined || widthPt !== undefined) {
-    const lnAttrs: Record<string, string | number> = {};
-    if (widthPt !== undefined) {
-      // OOXML stores stroke width in EMU (1 pt = 12 700 EMU). Round to
-      // the nearest integer because the schema types `w` as `xsd:int`.
-      lnAttrs.w = Math.round(widthPt * EMU_PER_PT);
-    }
-    const lnChildren: string[] = [];
-    if (fillHex) {
-      lnChildren.push(
-        xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: fillHex })]),
-      );
-    }
-    if (dash !== undefined) {
-      lnChildren.push(xmlSelfClose("a:prstDash", { val: dash }));
-    }
-    spPrChildren.push(
-      lnChildren.length === 0
-        ? xmlSelfClose("a:ln", lnAttrs)
-        : xmlElement("a:ln", Object.keys(lnAttrs).length > 0 ? lnAttrs : undefined, lnChildren),
-    );
-  }
-
-  return xmlElement("c:spPr", undefined, spPrChildren);
-}
-
-// ── Marker ───────────────────────────────────────────────────────────
-
-const VALID_MARKER_SYMBOLS: ReadonlySet<ChartMarkerSymbol> = new Set([
-  "none",
-  "auto",
-  "circle",
-  "square",
-  "diamond",
-  "triangle",
-  "x",
-  "star",
-  "dot",
-  "dash",
-  "plus",
-]);
-
-const MARKER_SIZE_MIN = 2;
-const MARKER_SIZE_MAX = 72;
-
-/**
- * Normalize a marker size to the OOXML 2..72 band (`ST_MarkerSize`).
- * Excel's UI clamps anything outside this range; we mirror that on the
- * write side so an out-of-range hint never reaches the chart XML.
- *
- * Returns `undefined` for non-finite values so the writer can elide
- * `<c:size>` (Excel falls back to its series-rotation default).
- */
-function clampMarkerSize(value: number | undefined): number | undefined {
-  if (value === undefined || !Number.isFinite(value)) return undefined;
-  const rounded = Math.round(value);
-  if (rounded < MARKER_SIZE_MIN) return MARKER_SIZE_MIN;
-  if (rounded > MARKER_SIZE_MAX) return MARKER_SIZE_MAX;
-  return rounded;
-}
-
-/**
- * Validate a marker symbol against the OOXML `ST_MarkerStyle` enum.
- * Returns `undefined` for unrecognized values so the writer can elide
- * `<c:symbol>` rather than emit a token Excel will reject.
- */
-function normalizeMarkerSymbol(
-  value: ChartMarkerSymbol | undefined,
-): ChartMarkerSymbol | undefined {
-  if (value === undefined) return undefined;
-  return VALID_MARKER_SYMBOLS.has(value) ? value : undefined;
-}
-
-/**
- * Build a `<c:marker>` element for a series. Returns `undefined` when
- * the marker block carries no meaningful settings — an empty marker
- * element collapses to the inherited series-rotation default Excel
- * picks anyway, so omitting it keeps untouched XML byte-clean.
- */
-function buildSeriesMarker(marker: ChartMarker | undefined): string | undefined {
-  if (!marker) return undefined;
-  const symbol = normalizeMarkerSymbol(marker.symbol);
-  const size = clampMarkerSize(marker.size);
-  const fill = normalizeRgbHex(marker.fill);
-  const line = normalizeRgbHex(marker.line);
-
-  if (symbol === undefined && size === undefined && !fill && !line) return undefined;
-
-  const children: string[] = [];
-  if (symbol !== undefined) children.push(xmlSelfClose("c:symbol", { val: symbol }));
-  if (size !== undefined) children.push(xmlSelfClose("c:size", { val: size }));
-
-  if (fill || line) {
-    const spPrChildren: string[] = [];
-    if (fill) {
-      spPrChildren.push(
-        xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: fill })]),
-      );
-    }
-    if (line) {
-      spPrChildren.push(
-        xmlElement("a:ln", undefined, [
-          xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: line })]),
-        ]),
-      );
-    }
-    children.push(xmlElement("c:spPr", undefined, spPrChildren));
-  }
-
-  return xmlElement("c:marker", undefined, children);
-}
-
-/**
- * Validate a 6-digit RGB hex string and return the uppercase form.
- * Strips a leading `#`. Returns `undefined` for missing or malformed
- * values so the writer can elide colored sub-elements that would
- * otherwise carry a token Excel rejects.
- */
-function normalizeRgbHex(value: string | undefined): string | undefined {
-  if (typeof value !== "string") return undefined;
-  const stripped = value.replace(/^#/, "").toUpperCase();
-  return /^[0-9A-F]{6}$/.test(stripped) ? stripped : undefined;
 }
 
 // ── Data Labels ──────────────────────────────────────────────────────
